% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_probability.R
\name{alert_probability}
\alias{alert_probability}
\title{Bayesian Alert Probability}
\usage{
alert_probability(posterior_probability, threshold = 0.95)
}
\arguments{
\item{posterior_probability}{Numeric vector of posterior probabilities.}

\item{threshold}{Alert threshold.}
}
\value{
Logical vector indicating alerts.
}
\description{
Bayesian Alert Probability
}
