% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor_alert.R
\name{bayes_factor_alert}
\alias{bayes_factor_alert}
\title{Approximate Bayes Factor Alert}
\usage{
bayes_factor_alert(posterior_prob, prior_prob = 0.5, threshold = 3)
}
\arguments{
\item{posterior_prob}{Posterior probability of deterioration.}

\item{prior_prob}{Prior probability of deterioration.}

\item{threshold}{Bayes factor threshold.}
}
\value{
Data frame with Bayes factor and alert.
}
\description{
Approximate Bayes Factor Alert
}
