% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_funnel_plot.R
\name{bayesian_funnel_plot}
\alias{bayesian_funnel_plot}
\title{Bayesian Funnel Plot Data}
\usage{
bayesian_funnel_plot(observed, expected, level = 0.95)
}
\arguments{
\item{observed}{Observed events.}

\item{expected}{Expected events.}

\item{level}{Credible interval level.}
}
\value{
A data frame for funnel plot construction.
}
\description{
Bayesian Funnel Plot Data
}
