% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_events.R
\name{expected_events}
\alias{expected_events}
\title{Calculate Expected Events}
\usage{
expected_events(risk, group = NULL)
}
\arguments{
\item{risk}{Predicted patient-level risks.}

\item{group}{Optional grouping variable.}
}
\value{
Expected event totals.
}
\description{
Calculate Expected Events
}
