% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o_minus_e.R
\name{o_minus_e}
\alias{o_minus_e}
\title{Observed Minus Expected Statistic}
\usage{
o_minus_e(observed, expected)
}
\arguments{
\item{observed}{Observed outcomes.}

\item{expected}{Expected outcomes.}
}
\value{
Data frame with O-E statistics.
}
\description{
Observed Minus Expected Statistic
}
