% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surveillance.R
\name{plot_surveillance}
\alias{plot_surveillance}
\title{Plot Surveillance Results}
\usage{
plot_surveillance(data, y = "posterior_mean")
}
\arguments{
\item{data}{Data frame returned by a surveillance function.}

\item{y}{Column name to plot.}
}
\value{
A ggplot object.
}
\description{
Plot Surveillance Results
}
