% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive.R
\name{posterior_predictive}
\alias{posterior_predictive}
\title{Posterior Predictive Probability}
\usage{
posterior_predictive(
  observed,
  predicted,
  sd_pred,
  direction = c("higher", "lower")
)
}
\arguments{
\item{observed}{Observed value.}

\item{predicted}{Predicted or expected value.}

\item{sd_pred}{Predictive standard deviation.}

\item{direction}{Direction of concern: "higher" or "lower".}
}
\value{
Posterior predictive tail probability.
}
\description{
Posterior Predictive Probability
}
