% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_adjusted_cusum.R
\name{risk_adjusted_cusum}
\alias{risk_adjusted_cusum}
\title{Risk-Adjusted CUSUM}
\usage{
risk_adjusted_cusum(observed, expected, threshold = 5)
}
\arguments{
\item{observed}{Observed outcomes.}

\item{expected}{Expected outcomes.}

\item{threshold}{Alert threshold.}
}
\value{
Risk-adjusted CUSUM results.
}
\description{
Risk-Adjusted CUSUM
}
