% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_adjusted_ewma.R
\name{risk_adjusted_ewma}
\alias{risk_adjusted_ewma}
\title{Risk-Adjusted EWMA}
\usage{
risk_adjusted_ewma(observed, expected, lambda = 0.2, threshold = 3)
}
\arguments{
\item{observed}{Observed outcomes.}

\item{expected}{Expected outcomes.}

\item{lambda}{EWMA smoothing parameter.}

\item{threshold}{Alert threshold.}
}
\value{
Risk-adjusted EWMA results.
}
\description{
Risk-Adjusted EWMA
}
