% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_surveillance.R
\name{roc_surveillance}
\alias{roc_surveillance}
\title{ROC Evaluation for Surveillance Alerts}
\usage{
roc_surveillance(truth, score)
}
\arguments{
\item{truth}{Binary true outcome.}

\item{score}{Numeric risk or alert score.}
}
\value{
A list containing ROC object and AUC.
}
\description{
ROC Evaluation for Surveillance Alerts
}
