% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BayesSurveillance.R
\name{summary.BayesSurveillance}
\alias{summary.BayesSurveillance}
\title{Summarise BayesSurveillance Results}
\usage{
\method{summary}{BayesSurveillance}(object, ...)
}
\arguments{
\item{object}{A surveillance result data frame.}

\item{...}{Additional arguments.}
}
\value{
Summary information.
}
\description{
Summarise BayesSurveillance Results
}
